+++ src/components/PopupModal/ModalContent.tsx
@@ -13,17 +13,11 @@ export interface Props {
   url: string;
   prefill?: Prefill;
   utm?: Utm;
-  styles?: React.CSSProperties | undefined;
   pageSettings?: PageSettings;
   iframeTitle?: IframeTitle;
 }
 
-const defaultStyles = {
-  minWidth: "320px",
-  height: "630px",
-};
-
-class InlineWidget extends React.Component<Props, { isLoading: boolean }> {
+class ModalContent extends React.Component<Props, { isLoading: boolean }> {
   constructor(props: Props) {
     super(props);
 
@@ -46,14 +40,11 @@ class InlineWidget extends React.Component<Props, { isLoading: boolean }> {
       pageSettings: this.props.pageSettings,
       prefill: this.props.prefill,
       utm: this.props.utm,
-      embedType: "Inline",
+      embedType: "PopupWidget",
     });
 
     return (
-      <div
-        className="calendly-inline-widget"
-        style={this.props.styles || defaultStyles}
-      >
+      <>
         {this.state.isLoading && <LoadingSpinner />}
         <iframe
           width="100%"
@@ -63,9 +54,9 @@ class InlineWidget extends React.Component<Props, { isLoading: boolean }> {
           onLoad={this.onLoad}
           src={src}
         ></iframe>
-      </div>
+      </>
     );
   }
 }
 
-export default InlineWidget;
+export default ModalContent;