+++ src/index.tsx
@@ -1,6 +1,7 @@
 import InlineWidget from "./components/InlineWidget/InlineWidget";
 import PopupButton from "./components/PopupButton/PopupButton";
 import PopupWidget from "./components/PopupWidget/PopupWidget";
+import PopupModal from "./components/PopupModal/Modal";
 import CalendlyEventListener from "./components/CalendlyEventListener/CalendlyEventListener";
 import type {
   DateAndTimeSelectedEvent,
@@ -8,11 +9,11 @@ import type {
   EventTypeViewedEvent,
   ProfilePageViewedEvent,
 } from "./components/CalendlyEventListener/CalendlyEventListener";
-import { openPopupWidget, closePopupWidget } from './calendly';
 
 export { InlineWidget };
 export { PopupButton };
 export { PopupWidget };
+export { PopupModal };
 export { CalendlyEventListener };
 export {
   DateAndTimeSelectedEvent,
@@ -20,4 +21,3 @@ export {
   EventTypeViewedEvent,
   ProfilePageViewedEvent,
 };
-export { openPopupWidget, closePopupWidget };